Add tests and fix multiple replacements.
This adds some tests in the form of examples of what replacements should look
like and a test runner that checks the actual output against the expected.

The tests identified an issue with processing where multiple excludes
or includes would be treated as one large include or exclude. This could
cause exclusion of lines which should be included and the other way around.
This was due to using too greedy of a regex and the issue was fixed.
Also, the handling of newlines was refined.

Fixes #1.